refactored the entire project
