+++ src/game/player.rs
@@ -1,5 +1,7 @@
-use crate::utils::{collide, Collision};
 use bevy::prelude::*;
+
+use crate::physics::collide;
+use crate::physics::Collision;
 
 use super::Collider;
 use super::ColliderKind;
@@ -39,12 +41,12 @@ fn setup(
     mut commands: Commands,
     asset_server: Res<AssetServer>,
     mut materials: ResMut<Assets<ColorMaterial>>,
-    mut texture_atlases: ResMut<Assets<TextureAtlas>>,
+    mut _texture_atlases: ResMut<Assets<TextureAtlas>>,
 ) {
     // player
     let texture_handle = asset_server.load("textures/rpg/chars/example.png");
-    let texture_atlas = TextureAtlas::from_grid(texture_handle, Vec2::new(100.0, 100.0), 10, 10);
-    let texture_atlas_handle = texture_atlases.add(texture_atlas);
+    let _texture_atlas = TextureAtlas::from_grid(texture_handle, Vec2::new(100.0, 100.0), 10, 10);
+    // let texture_atlas_handle = texture_atlases.add(texture_atlas);
 
     //commands
     //    .spawn_bundle(SpriteSheetBundle {
