+++ src/game/spikes.rs
@@ -122,7 +122,6 @@ fn generate_new_spikes(mut commands: Commands, mut materials: ResMut<Assets<Colo
         &mut commands,
         material,
         position,
-        Vec2::new(0.0, 0.0),
         size,
         10.0,
         direction,
@@ -155,7 +154,6 @@ fn spawn_spike(
     commands: &mut Commands,
     material: &Handle<ColorMaterial>,
     position: Vec2,
-    rotation: Vec2,
     size: Vec2,
     thickness: f32,
     direction: Direction,
