+++ src/gamedata.rs
@@ -1,6 +1,7 @@
 use bevy::prelude::*;
 
 pub struct GameDataPlugin;
+
 impl Plugin for GameDataPlugin {
     fn build(&self, app: &mut AppBuilder) {
         app.insert_resource(Score { score: -1 });
